requirements.txt
-e git+https://github.com/hedleyroos/django-parler@master#egg=django_parler
asgiref==3.4.1
Django==3.2.6
django-photologue==3.13
django-sortedm2m==3.0.2
ExifRead==2.3.2
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
